Comment out main(…) in non-essential methods
Because we are using main as entry points, we have to be careful not to
pollute our analysis with spurious main(…) that have been inserted as a
quick way to do testing by the original authors.

Reducing the number of mains (i.e., the entry points) makes the analysis
faster.